requirements.txt
streamlit
google-genai
python-dotenv
pillow
watchdog
